feat(Monitoring): outsource tests to the files they belong
